:green_heart: Fikset tester for token CI
